refactored fab menu
