+++ scripts/utils/mobileView.js
@@ -1,0 +1,3 @@
+export function isMobileView() {
+  return window.matchMedia('(max-width: 768px)').matches
+}
